Changed column type to decimal

+++ db/migrate/20131109165332_create_books.rb
@@ -1,7 +1,7 @@
 class CreateBooks < ActiveRecord::Migration
   def change
     create_table :books do |t|
-      t.integer :isbn
+      t.decimal :isbn
       t.string :title
       t.string :image_url
       t.decimal :price, precision:8, scale: 2
